requirements.txt
streamlit
pytest
pytest-mock
langchain
langchain-openai
langchain-core
openai
anthropic
python-dotenv
plotly
fpdf2
